+++ InedoCore/InedoExtension/UserDirectories/Clients/LdapClientEntry.cs
